Add auto_smooth_up and auto_smooth_down scrolling
This adds the auto_smooth_up() and auto_smooth_down() scrolling
functions to the Canvas, which only scroll smooth if the current layout
supports smooth scrolling. Otherwise fall back to page_up/page_down
scrolling.

Keys J and K are now bound to that instead of to the old smooth_down and
smooth_up.