Add basic dbus-API for synctex support
Katarakt now exposes a dbus-API that allows TeX-editors to synchronize
the positions in source code and pdf via synctex.

TODO: The API allows to specify positions in the PDF as (page,x,y)
triple, but the x,y coordinates are currently ignored or set to 0,
respectively. We need more katarakt-internal API for this to be
implementable.